chore: move exported types back to index.ts
